Refactor exceptions into seperate file and add base class
